+++ example/src/demos/Lines.tsx
@@ -1,5 +1,5 @@
 import React, { useRef, useEffect, useState, useCallback, useContext, useMemo } from 'react'
-import { extend, Canvas, useThree, ReactThreeFiber } from '@react-three/fiber'
+import { extend, Canvas, useThree, ReactThreeFiber, ThreeEvent } from '@react-three/fiber'
 import { OrbitControls } from 'three-stdlib'
 extend({ OrbitControls })
 
@@ -27,43 +27,33 @@ function useDrag(onDrag: any, onEnd: any) {
   const [active, setActive] = useState(false)
   const [, toggle] = useContext(camContext) as any
 
-  const down = useCallback(
-    (e) => {
-      console.log('down')
-      setActive(true)
-      toggle(false)
-      e.stopPropagation()
-      e.target.setPointerCapture(e.pointerId)
-    },
-    [toggle],
-  )
+  const down = (event: ThreeEvent<PointerEvent>) => {
+    console.log('down')
+    setActive(true)
+    toggle(false)
+    event.stopPropagation()
+    // @ts-ignore
+    event.target.setPointerCapture(event.pointerId)
+  }
 
-  const up = useCallback(
-    (e) => {
-      console.log('up')
-      setActive(false)
-      toggle(true)
-      e.stopPropagation()
-      e.target.releasePointerCapture(e.pointerId)
-      if (onEnd) onEnd()
-    },
-    [onEnd, toggle],
-  )
+  const up = (event: ThreeEvent<PointerEvent>) => {
+    console.log('up')
+    setActive(false)
+    toggle(true)
+    event.stopPropagation()
+    // @ts-ignore
+    event.target.releasePointerCapture(event.pointerId)
+    if (onEnd) onEnd()
+  }
 
-  const activeRef = useRef<any>()
-  useEffect(() => void (activeRef.current = active))
-  const move = useCallback(
-    (event) => {
-      if (activeRef.current) {
-        event.stopPropagation()
-        onDrag(event.unprojectedPoint)
-      }
-    },
-    [onDrag],
-  )
+  const move = (event: ThreeEvent<PointerEvent>) => {
+    if (active) {
+      event.stopPropagation()
+      onDrag(event.unprojectedPoint)
+    }
+  }
 
-  const [bind] = useState(() => ({ onPointerDown: down, onPointerUp: up, onPointerMove: move }))
-  return bind
+  return { onPointerDown: down, onPointerUp: up, onPointerMove: move }
 }
 
 function EndPoint({ position, onDrag, onEnd }: any) {