+++ example/src/demos/StopPropagation.tsx
@@ -5,9 +5,7 @@ import { useGLTF, Environment, OrbitControls } from '@react-three/drei'
 
 function Soda(props: any) {
   const [hovered, spread] = useHover()
-  const { nodes, materials } = useGLTF(
-    'https://market-assets.fra1.cdn.digitaloceanspaces.com/market-assets/models/soda-bottle/model.gltf',
-  ) as any
+  const { nodes, materials } = useGLTF('/bottle.gltf') as any
   return (
     <group {...props} {...spread} dispose={null}>
       <mesh geometry={nodes.Mesh_sodaBottle.geometry}>