+++ packages/fiber/src/native/Canvas.tsx
@@ -9,10 +9,12 @@ import { createTouchEvents } from './events'
 import { RootState, Size } from '../core/store'
 import { polyfills } from './polyfills'
 
-export interface Props extends Omit<RenderProps<HTMLCanvasElement>, 'size' | 'dpr'>, ViewProps {
+export interface CanvasProps extends Omit<RenderProps<HTMLCanvasElement>, 'size' | 'dpr'>, ViewProps {
   children: React.ReactNode
   style?: ViewStyle
 }
+
+export interface Props extends CanvasProps {}
 
 /**
  * A native canvas which accepts threejs elements as children.
@@ -62,7 +64,7 @@ const CanvasImpl = /*#__PURE__*/ React.forwardRef<View, Props>(
     if (error) throw error
 
     const viewRef = React.useRef<View>(null!)
-    const root = React.useRef<ReconcilerRoot<Element>>(null!)
+    const root = React.useRef<ReconcilerRoot<HTMLCanvasElement>>(null!)
 
     // Inject and cleanup RN polyfills if able
     React.useLayoutEffect(() => polyfills(), [])
@@ -85,7 +87,7 @@ const CanvasImpl = /*#__PURE__*/ React.forwardRef<View, Props>(
         getContext: (() => context) as any,
       } as HTMLCanvasElement
 
-      root.current = createRoot<Element>(canvasShim)
+      root.current = createRoot<HTMLCanvasElement>(canvasShim)
       setCanvas(canvasShim)
     }, [])